change the method of obtaining the book image, add display of books from the database on the page

+++ src/main/java/ru/alex/bookstore/dto/BookCreateEditDto.java
@@ -5,7 +5,7 @@ import org.springframework.web.multipart.MultipartFile;
 public record BookCreateEditDto(
         String name,
         String author,
-        Integer rating,
+        Float rating,
         Integer price,
         MultipartFile image,
         String publisher,
